refactor: 각 이벤트 인터페이스 도출
