+++ src/main/java/cart/application/event/MemberPointUpdateEvent.java
@@ -3,31 +3,11 @@ package cart.application.event;
 import cart.domain.member.Member;
 import cart.dto.request.OrderRequest;
 
-public class MemberPointUpdateEvent {
+public interface MemberPointUpdateEvent {
 
-    private final Member member;
-    private final OrderRequest orderRequest;
-    private final Long orderId;
+    Member getMember();
 
-    public MemberPointUpdateEvent(
-            final Member member,
-            final OrderRequest orderRequest,
-            final Long orderId
-    ) {
-        this.member = member;
-        this.orderRequest = orderRequest;
-        this.orderId = orderId;
-    }
+    OrderRequest getOrderRequest();
 
-    public Member getMember() {
-        return member;
-    }
-
-    public OrderRequest getOrderRequest() {
-        return orderRequest;
-    }
-
-    public Long getOrderId() {
-        return orderId;
-    }
+    Long getOrderId();
 }
