+++ src/main/java/cart/application/event/PaymentRequestEvent.java
@@ -4,32 +4,11 @@ import cart.domain.member.Member;
 import cart.domain.product.Price;
 import cart.dto.request.OrderRequest;
 
-public class PaymentRequestEvent {
+public interface PaymentRequestEvent {
 
-    private final Member member;
-    private final Price price;
-    private final OrderRequest orderRequest;
+    Member getMember();
 
-    public PaymentRequestEvent(
-            final Member member,
-            final Price price,
-            final OrderRequest orderRequest
+    Price getPrice();
 
-    ) {
-        this.member = member;
-        this.price = price;
-        this.orderRequest = orderRequest;
-    }
-
-    public Member getMember() {
-        return member;
-    }
-
-    public Price getPrice() {
-        return price;
-    }
-
-    public OrderRequest getOrderRequest() {
-        return orderRequest;
-    }
+    OrderRequest getOrderRequest();
 }
