+++ src/test/java/cart/application/PointServiceTest.java
@@ -5,7 +5,7 @@ import static org.junit.jupiter.api.Assertions.assertAll;
 import static org.mockito.BDDMockito.given;
 import static org.mockito.Mockito.verify;
 
-import cart.application.event.MemberPointUpdateEvent;
+import cart.application.event.DefaultMemberPointUpdateEvent;
 import cart.domain.cart.Quantity;
 import cart.domain.member.Email;
 import cart.domain.member.Member;
@@ -72,7 +72,7 @@ class PointServiceTest {
         given(orderRepository.findByOrderId(1L)).willReturn(order);
 
         // when
-        pointService.updateMemberPoint(new MemberPointUpdateEvent(member, orderRequest, 1L));
+        pointService.updateMemberPoint(new DefaultMemberPointUpdateEvent(member, orderRequest, 1L));
 
         //then
         ArgumentCaptor<Member> memberArgumentCaptor = ArgumentCaptor.forClass(Member.class);
@@ -106,7 +106,7 @@ class PointServiceTest {
         given(orderRepository.findByOrderId(1L)).willReturn(order);
 
         // when
-        pointService.updateMemberPoint(new MemberPointUpdateEvent(member, orderRequest, 1L));
+        pointService.updateMemberPoint(new DefaultMemberPointUpdateEvent(member, orderRequest, 1L));
 
         //then
         ArgumentCaptor<Order> orderArgumentCaptor = ArgumentCaptor.forClass(Order.class);
